Changes to NavBar and url access relating to admin priveleges
