+++ Tabloid/client/src/App.js
@@ -4,18 +4,32 @@ import { Spinner } from "reactstrap";
 import Header from "./components/Header";
 import ApplicationViews from "./components/ApplicationViews";
 import { onLoginStatusChange } from "./modules/authManager";
-
+import { getCurrentUserType } from "./modules/userManager";
 
 function App() {
   const [isLoggedIn, setIsLoggedIn] = useState(null);
- 
 
-  
+  const [isAdmin, setIsAdmin] = useState(false);
+
+  const userIsAdmin = () => {
+    getCurrentUserType().then((userType) => {
+      if (userType.name === "Admin") {
+        setIsAdmin(true);
+      } else {
+        setIsAdmin(false);
+      }
+    });
+  };
 
   useEffect(() => {
     onLoginStatusChange(setIsLoggedIn);
-   
   }, []);
+
+  useEffect(() => {
+    if (isLoggedIn) {
+      userIsAdmin();
+    }
+  }, [isLoggedIn]);
 
   if (isLoggedIn === null) {
     return <Spinner className="app-spinner dark" />;
@@ -23,8 +37,8 @@ function App() {
 
   return (
     <Router>
-      <Header isLoggedIn={isLoggedIn} />
-      <ApplicationViews isLoggedIn={isLoggedIn} />
+      <Header isLoggedIn={isLoggedIn} isAdmin={isAdmin} />
+      <ApplicationViews isLoggedIn={isLoggedIn} isAdmin={isAdmin} />
     </Router>
   );
 }
