+++ Tabloid/client/src/components/ApplicationViews.js
@@ -16,7 +16,7 @@ import PostDetails from "./PostDetail";
 import UserDetails from "./UserDetails";
 import PostForm from "./PostForm";
 
-export default function ApplicationViews({ isLoggedIn }) {
+export default function ApplicationViews({ isLoggedIn, isAdmin }) {
   return (
     <main>
       <Switch>
@@ -25,11 +25,21 @@ export default function ApplicationViews({ isLoggedIn }) {
         </Route>
 
         <Route exact path="/tags">
-          {isLoggedIn ? <TagList /> : <Redirect to="/login" />}
+          {/* {isLoggedIn ? <TagList /> : <Redirect to="/login" />} */}
+          {isLoggedIn && isAdmin ? (
+            <TagList />
+          ) : (
+            [isLoggedIn && !isAdmin ? <Hello /> : <Redirect to="/login" />]
+          )}
         </Route>
 
         <Route exact path="/tags/add">
-          {isLoggedIn ? <TagForm /> : <Redirect to="/login" />}
+          {/* {isLoggedIn ? <TagForm /> : <Redirect to="/login" />} */}
+          {isLoggedIn && isAdmin ? (
+            <TagForm />
+          ) : (
+            [isLoggedIn && !isAdmin ? <Hello /> : <Redirect to="/login" />]
+          )}
         </Route>
 
         <Route path="/login">
@@ -41,11 +51,21 @@ export default function ApplicationViews({ isLoggedIn }) {
         </Route>
 
         <Route exact path="/category">
-          {isLoggedIn ? <CategoryList /> : <Redirect to="/login" />}
+          {/* {isLoggedIn ? <CategoryList /> : <Redirect to="/login" />} */}
+          {isLoggedIn && isAdmin ? (
+            <CategoryList />
+          ) : (
+            [isLoggedIn && !isAdmin ? <Hello /> : <Redirect to="/login" />]
+          )}
         </Route>
 
         <Route path="/category/add">
-          <CategoryForm />
+          {/* <CategoryForm /> */}
+          {isLoggedIn && isAdmin ? (
+            <CategoryForm />
+          ) : (
+            [isLoggedIn && !isAdmin ? <Hello /> : <Redirect to="/login" />]
+          )}
         </Route>
 
         <Route exact path="/post">
@@ -58,7 +78,7 @@ export default function ApplicationViews({ isLoggedIn }) {
           <CommentAddForm />
         </Route>
         <Route exact path="/myPost">
-          <MyPost /> 
+          <MyPost />
         </Route>
         <Route path="/post/add">
           {isLoggedIn ? <PostForm /> : <Redirect to="/login" />}
@@ -68,10 +88,18 @@ export default function ApplicationViews({ isLoggedIn }) {
           {isLoggedIn ? <PostDetails /> : <Redirect to="/login" />}
         </Route>
         <Route exact path="/users">
-          {isLoggedIn ? <UserList /> : <Redirect to="/login" />}
+          {isLoggedIn && isAdmin ? (
+            <UserList />
+          ) : (
+            [isLoggedIn && !isAdmin ? <Hello /> : <Redirect to="/login" />]
+          )}
         </Route>
         <Route path="/users/:id(\d+)">
-          {isLoggedIn ? <UserDetails /> : <Redirect to="/login" />}
+          {isLoggedIn && isAdmin ? (
+            <UserDetails />
+          ) : (
+            [isLoggedIn && !isAdmin ? <Hello /> : <Redirect to="/login" />]
+          )}
         </Route>
       </Switch>
     </main>
